requirements.txt
flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
flask-login==0.6.3
flask-cors==3.0.10
sentence-transformers==2.2.2
pinecone-client==3.0.0
huggingface_hub==0.13.4
torch==1.13.1
transformers==4.28.1
werkzeug==2.3.7
email-validator==2.1.0.post1
